Support delayed stack trace loading
This capability of the Debug Adapter Protocol supports
the delayed loading of parts of the stack. It requires
that both the 'startFrame' and 'levels' arguments and
the 'totalFrames' result of the 'StackTrace' requests
are supported. It helps to avoid duplicated top stack
frame and limited number of frames in the CALLSTACK
pane (how many can be transmitted in one message).

IoT.js-VSCode-DCO-1.0-Signed-off-by: Daniella Barsony bella@inf.u-szeged.hu
IoT.js-VSCode-DCO-1.0-Signed-off-by: Robert Sipka rsipka.uszeged@partner.samsung.com