Highlight where we stopped in the source code
The ScopesRequest will called automatically after the first StackTraceResponse.
It is responsible for highlighting where we are in the code.
Since the backtrace array was empty after the first response, it was not called properly.
Handling the BACTRACE_TOTAL response message separately resolves this issue.

IoT.js-VSCode-DCO-1.0-Signed-off-by: Robert Sipka rsipka.uszeged@partner.samsung.com